fix: force float division

+++ src/game/rect.cpp
@@ -108,7 +108,7 @@ Rect Rect::getVerticalSlice(unsigned int ratio, unsigned int index) const {
     Rect slice{*this};
 
     // A "unit width" conforming to a common ratio.
-    slice.w = std::round(this->w / ratio);
+    slice.w = std::round((float)this->w / ratio);
 
     // A "unit distance" from the left edge of the rect.
     slice.x += (index * slice.w);
